requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
blas=1.0=mkl
ca-certificates=2021.1.19=h06a4308_0
certifi=2020.12.5=py38h06a4308_0
cudatoolkit=11.0.221=h6bb024c_0
freetype=2.10.4=h5ab3b9f_0
intel-openmp=2020.2=254
joblib=1.0.1=pyhd3eb1b0_0
jpeg=9b=h024ee3a_2
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
libuv=1.40.0=h7b6447c_0
lz4-c=1.9.3=h2531618_0
mkl=2020.2=256
mkl-service=2.3.0=py38he904b0f_0
mkl_fft=1.3.0=py38h54f3939_0
mkl_random=1.1.1=py38h0573a6f_0
ncurses=6.2=he6710b0_1
ninja=1.10.2=py38hff7bd54_0
numpy=1.19.2=py38h54aff64_0
numpy-base=1.19.2=py38hfa32c7d_0
olefile=0.46=py_0
openssl=1.1.1j=h27cfd23_0
pandas=1.2.2=py38ha9443f7_0
pillow=8.1.0=py38he98fc37_0
pip=21.0.1=py38h06a4308_0
python=3.8.8=hdb3f193_4
python-dateutil=2.8.1=pyhd3eb1b0_0
pytorch=1.7.1=py3.8_cuda11.0.221_cudnn8.0.5_0
pytz=2021.1=pyhd3eb1b0_0
readline=8.1=h27cfd23_0
scikit-learn=0.24.1=py38ha9443f7_0
scipy=1.6.1=py38h91f5cce_0
setuptools=52.0.0=py38h06a4308_0
six=1.15.0=py38h06a4308_0
sqlite=3.33.0=h62c20be_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=hbc83047_0
torchaudio=0.7.2=py38
torchvision=0.8.2=py38_cu110
typing_extensions=3.7.4.3=pyha847dfd_0
wheel=0.36.2=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.5=h9ceee32_0
